Added trust config
The trustConfig is configurable via a different file. Location is in the data_dir/trusted_issuers.yaml
There's also an API to add trust/remove trust which also saves the file. This comes in handy for testing and on-the-fly adding of trust.
It's exptected that a production network doesn't need these but only uses the file.