refactor(crypto): use jwt.NewBuilder() instead of jwt.New()
Proposal in order to evaluate if this (builder) way of generating a
JWT is cleaner.

Initial impression is that this way of generating a JWT does not yield
a big benefit. It possibly introduces inconsistent way's of generating
a JWT. Actual builder like method chaining is limited given the claims
map it has to parse.

There are similar calls in oauth_test.go that can be refactored. These
have been ignored given the outcome of this proposal.

Refs: #665