Enhance key resolution in DID documents
The commit improves the key resolution process of DID documents in the 'resolver' package. The document can define a base by setting the @context/@base value and refer to keys from their hash value, so did:example#id is defined as #id if the @context/@base value is set to did:example.