Add conditional for storageClassName in nuts-data-pv.yaml (#2750)
Incorporated a conditional that checks if 'storageClassName' exists under .Values.nuts.data.persistedVolume in nuts-data-pv.yaml. The GKE platform sets the storageClassName to its own value after creation. When Helm runs a diff on the upgrade command, the change of the value of the storageClassName is detected. Helm than tries to patch the PVC, causing an error as it is immutable. This causes the upgrade command to fail. This solution only sets the storageClassName if the values.yaml explicitly does so, preventing this error condition to happen.