Add Delete function to vcr wallet
A new Delete function is added to the wallet in the vcr/holder/wallet.go file. This function allows a credential to be removed from the wallet. Additionally, an interface in the vcr/holder/interface.go file has been updated to include this Delete function. Proper error handling and stats consistency are ensured in this implementation.