Added endpoints for Oid4VCI flow and operations for wallet in auth IAM API
Implemented capability for the initiation and completion of the Oid4VCI flow in the auth IAM API, allowing Credential Issuers to request and validate credentials. Additionally, endpoints for operations such as search and delete are provided for managing wallet credentials. These changes are expected to enhance the user journey for issuing and managing verifiable credentials.