Refactor VerifiableCredentials function to include nonce
The VerifiableCredentials function has been revised to involve the nonce from token response in IAM authentication process. The nonce, which is used in the JWT proof of possession, was previously not included in the VerifiableCredentials function. Necessary updates were also made in related test cases to pass the appropriate nonce.